Creating Qt4_TangoLimaVideo Modify Qt4_LimaVideo so that both rely on a common GenericVideoDevice.py device object
